% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slurm_utils.R
\name{cleanup_files}
\alias{cleanup_files}
\title{Deletes temporary files associated with a SLURM job}
\usage{
cleanup_files(slr_job)
}
\arguments{
\item{slr_job}{A \code{slurm_job} object output by \code{\link{slurm_apply}}.}
}
\description{
This function deletes all temporary files associated with the specified SLURM
job, including files created by \code{\link{slurm_apply}} (parameters data, R
and Bash scripts, all starting with "slr####"), function output files
("slr####_[node_id].out") as well as the console and error output files
(starting with "slurm-[job_id]").
}
\examples{
\dontrun{
sjob <- slurm_apply(func, pars)
print_job_status(sjob) # Prints console/error output once job is completed.
func_result <- get_slurm_out(sjob, "table") # Loads output data into R.
cleanup_files(sjob)
}
}
\seealso{
\code{\link{slurm_apply}}, \code{\link{print_job_status}},
  \code{\link{get_slurm_out}}
}

