% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomkill_handling.R
\name{findIntRuns}
\alias{findIntRuns}
\title{Collapse runs of consecutive numbers Borrowed from}
\usage{
findIntRuns(run)
}
\arguments{
\item{run}{A sequence of integers.}
}
\value{
A sequence of integers with collapsed consecutive runs.
}
\description{
Collapse runs of consecutive numbers Borrowed from
}
