% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomkill_handling.R
\name{get_oomKill_arrayindex}
\alias{get_oomKill_arrayindex}
\title{Find the array index of jobs killed due to out-of-memory fault.}
\usage{
get_oomKill_arrayindex(slr_job, wait)
}
\arguments{
\item{slr_job}{A \code{slurm_job} object.}

\item{wait}{Specify whether to block until \code{slr_job} completes.}
}
\value{
A vector of integers corresponding to the array index of jobs killed
  due to out-of-memory fault.
}
\description{
Find the array index of jobs killed due to out-of-memory fault.
}
\examples{
\dontrun{sjob_oomkilled <- get_oomKill_arrayindex(sjob = sjob)}

}
