% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomkill_handling.R
\name{get_sbatch_script}
\alias{get_sbatch_script}
\title{Read an sbatch script from an rslurm array job}
\usage{
get_sbatch_script(slr_job, as_path = TRUE)
}
\arguments{
\item{slr_job}{A \code{slurm_job} object.}

\item{as_path}{Returns the path to the slurm array job script (T/F).}
}
\value{
The path to the slurm array job script (as_path = T) or a dataframe
  containing the slurm array job file inputs (as_path = F).
}
\description{
Read an sbatch script from an rslurm array job
}
\examples{
\dontrun{get_sbatch_script(slr_job = sjob, as_path = T)}

}
