% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rslurm.R
\name{print_job_status}
\alias{print_job_status}
\title{Print the status of a SLURM job and, if completed, its console/error output.}
\usage{
print_job_status(slr_job)
}
\arguments{
\item{slr_job}{A \code{slurm_job} object output by \code{\link{slurm_apply}}.}
}
\description{
If the specified SLURM job is still in the queue or running, this function
prints its current status (as output by the SLURM \code{squeue} command). If it
has completed or stopped, the function prints the console and error output,
if any, generated by each node.
}
\seealso{
\code{\link{slurm_apply}}
}

