\name{Slurm-package}
\alias{Slurm-package}
\docType{package}
\title{Slurm Job Environment}
\usage{
slurm.jobid()
slurm.array_job_id()
slurm.array_task_id()
slurm.arrayid()
slurm.job_name()
}
}
\description{
Provides Slurm job environment information.
}
\details{
\tabular{ll}{
Package: \tab Slurm\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-10-10\cr
License: \tab GPLv2\cr
}
}
\author{
Chris Scheller

Maintainer: Chris Scheller <schelcj@umich.edu>
}
\references{
\url{http://slurm.schedmd.com/}
}
\keyword{slurm}
%% list methods