\name{slurm.job_name}
\alias{slurm.job_name}
\title{
SLURM job name
}
\description{
SLURM job name as specified by \code{--job-name} or autogenerated if not specified.
} 
\usage{
slurm.job_name()
}
\details{
SLURM job name as specified by \code{--job-name} or autogenerated if not specified.
}
\value{
  \item{string}{Current value of SLURM_JOB_NAME}
}
\author{
Chris Scheller
}