% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slurm_call.R
\name{slurm_call}
\alias{slurm_call}
\title{Execution of a single function call on the SLURM cluster}
\usage{
slurm_call(f, params, data_file = NULL, pkgs = rev(.packages()))
}
\arguments{
\item{f}{Any R function.}

\item{params}{A named list of parameters to pass to \code{f}.}

\item{data_file}{The name of a R data file (created with
\code{\link[base]{save}}) to be loaded prior to calling \code{f}.}

\item{pkgs}{A character vector containing the names of packages that must
be loaded on each cluster node. By default, it includes all packages
loaded by the user when \code{slurm_call} is called.}
}
\value{
A \code{slurm_job} object containing the \code{file_prefix} assigned
  to temporary files created by \code{slurm_call}, a \code{job_id} assigned
  by the SLURM cluster, the number of \code{nodes} used (always 1 for
  \code{slurm_call}) and the \code{output} type (always 'raw').
}
\description{
Use \code{slurm_call} to perform a single function evaluation on the SLURM cluster.
}
\details{
This function creates temporary files for the parameters data ('slr####.RData'),
the R script evaluating the function ('slr####.R') and the Bash script
('slr####.sh') that sends the batch run command to the cluster.

Any other R objects (besides \code{params}) that \code{f} needs to access
should be saved in a .RData file (using \code{\link[base]{save}}) and the
name of this file should be given as the optional \code{data_file} argument.

When processing the computation job, the SLURM cluster will output two files:
one .RData file containing the function output ('slr####_0.RData') and one
containing any console or error output produced by R ('slurm-[job_id].out').

After sending the job to the SLURM cluster, \code{slurm_call} returns a
\code{slurm_job} object which can be used to cancel the job, get the job
status or output, and delete the temporary files associated with it. See
the description of the related functions for more details.
}
\seealso{
\code{\link{slurm_apply}} to parallelize a function over a parameter set.

\code{\link{cancel_slurm}}, \code{\link{cleanup_files}},
  \code{\link{get_slurm_out}} and \code{\link{print_job_status}}
  which use the output of this function.
}

