% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oomkill_handling.R
\name{update_sbatch_script}
\alias{update_sbatch_script}
\title{Update a slurm array job script}
\usage{
update_sbatch_script(slr_job = NA, array_index = NULL, mem = NULL,
  partition = NULL, write_script = FALSE, submit_job = FALSE)
}
\arguments{
\item{slr_job}{A \code{slurm_job} object.}

\item{array_index}{A vector of integers indicating array indicies to be submitted with the sbatch job.}

\item{mem}{Amount of memeory to request for the sbatch array.}

\item{partition}{The partition to request for the sbatch.}

\item{write_script}{Write the script to file (T/F).}

\item{submit_job}{Should the array job be submitted (T/F). Only possible when
\code{write_script = T}.}
}
\value{
Returns the updated script as a matrix or overwrites the existing
  slurm array job script.
}
\description{
Update a slurm array job script
}
\examples{
\dontrun{update_sbatch_script(slr_job = sjob, mem = 1, write_script = FALSE) }

}
