% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel_print_cleanup.R
\name{cancel_slurm}
\alias{cancel_slurm}
\title{Cancels a scheduled SLURM job}
\usage{
cancel_slurm(slr_job)
}
\arguments{
\item{slr_job}{A \code{slurm_job} object.}
}
\description{
This function cancels the specified SLURM job by invoking the SLURM 
\code{scancel} command. It does \emph{not} delete the temporary files 
(e.g. scripts) created by \code{\link{slurm_apply}} or 
\code{\link{slurm_call}}. Use \code{\link{cleanup_files}} to remove those 
files.
}
\seealso{
\code{\link{cleanup_files}}
}
