% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/slurm_utils.R
\name{get_slurm_out}
\alias{get_slurm_out}
\title{Reads the output of a function calculated on the SLURM cluster}
\usage{
get_slurm_out(slr_job)
}
\arguments{
\item{slr_job}{A \code{slurm_job} object output by \code{\link{slurm_apply}}.}
}
\value{
A data frame with one column by return value of the function passed
  to \code{\link{slurm_apply}}, where each row is the output of the
  corresponding row in the params data frame passed to
  \code{\link{slurm_apply}}.
}
\description{
This function reads all function output files (one by cluster node used) from
the specified SLURM job and returns the result in a single data frame. It
doesn't record any messages (including warnings or errors) output to the R
console during the computation; these can be consulted by invoking
\code{\link{print_job_status}}.
}
\seealso{
\code{\link{slurm_apply}}, \code{\link{print_job_status}}
}

